Refactor plugin to custom field example structure
