+++ lib/discourse_dice_comment/engine.rb
@@ -1,0 +1,6 @@
+module ::DiscourseDiceComment
+  class Engine < ::Rails::Engine
+    engine_name "discourse_dice_comment"
+    isolate_namespace DiscourseDiceComment
+  end
+end
