+++ plugin.rb
@@ -5,9 +5,22 @@
 # version: 0.2
 # authors: Codex
 
+enabled_site_setting :dice_comment_enabled
+
+register_topic_custom_field_type('dice_only', :boolean)
+register_topic_custom_field_type('dice_min', :integer)
+register_topic_custom_field_type('dice_max', :integer)
+
+register_asset 'stylesheets/common/dice-comment.scss'
+register_asset 'javascripts/discourse/initializers/add-dice-checkbox.js', :server_side
+register_asset 'javascripts/discourse/initializers/dice-button.js', :server_side
+
 after_initialize do
-  # Store dice fields when topic is created
-  on(:topic_created) do |topic, params, user|
+  require_relative 'lib/discourse_dice_comment/engine'
+  require_relative 'config/routes'
+  load File.expand_path('app/serializers/topic_view_serializer_extension.rb', __dir__)
+
+  DiscourseEvent.on(:topic_created) do |topic, params, _user|
     if params[:dice_only].present?
       topic.custom_fields['dice_only'] = params[:dice_only]
       topic.custom_fields['dice_min'] = params[:dice_min]
@@ -15,41 +28,4 @@ after_initialize do
       topic.save_custom_fields
     end
   end
-
-  module ::DiscourseDiceComment
-    class Engine < ::Rails::Engine
-      engine_name "discourse_dice_comment"
-      isolate_namespace DiscourseDiceComment
-    end
-  end
-
-  require_relative 'config/routes'
-
-  class ::DiscourseDiceComment::RollController < ::ApplicationController
-    requires_plugin 'discourse-dice-comment'
-    before_action :ensure_logged_in
-
-    def create
-      topic = Topic.find(params[:topic_id])
-      raise Discourse::NotFound unless topic.custom_fields['dice_only']
-
-      min = topic.custom_fields['dice_min'].to_i
-      max = topic.custom_fields['dice_max'].to_i
-      min = 0 if min < 0
-      max = 100 if max <= 0 || max < min
-
-      roll = rand(min..max)
-      raw = "\u{1F3B2} #{roll}! #{current_user.username}\uB2D8\uC758 \uC6B4\uBA85\uC740!?"
-      PostCreator.create!(current_user, topic_id: topic.id, raw: raw)
-
-      render json: success_json
-    end
-  end
-
-  add_to_serializer(:topic_view, :dice_only) { object.topic.custom_fields['dice_only'] }
-  add_to_serializer(:topic_view, :include_dice_only?) { object.topic.custom_fields.key?('dice_only') }
-  add_to_serializer(:topic_view, :dice_min) { object.topic.custom_fields['dice_min'] }
-  add_to_serializer(:topic_view, :include_dice_min?) { object.topic.custom_fields.key?('dice_min') }
-  add_to_serializer(:topic_view, :dice_max) { object.topic.custom_fields['dice_max'] }
-  add_to_serializer(:topic_view, :include_dice_max?) { object.topic.custom_fields.key?('dice_max') }
 end
